+++ src/matterbridge.ts
@@ -26,7 +26,8 @@ import { AnsiLogger, BRIGHT, RESET, TimestampFormat, UNDERLINE, UNDERLINEOFF, YE
 import { fileURLToPath, pathToFileURL } from 'url';
 import { promises as fs } from 'fs';
 import { ExecException, exec, spawn } from 'child_process';
-import { Server } from 'http';
+import { Server, createServer } from 'http';
+import * as http from 'http';
 import https from 'https';
 import EventEmitter from 'events';
 import express from 'express';
@@ -198,8 +199,11 @@ export class Matterbridge extends EventEmitter {
   private registeredDevices: RegisteredDevice[] = [];
   private nodeStorage: NodeStorageManager | undefined;
   private nodeContext: NodeStorage | undefined;
+
   private expressApp: express.Express | undefined;
   private expressServer: Server | undefined;
+  private httpServer: Server | undefined;
+  private httpsServer: Server | undefined;
   private webSocketServer: WebSocketServer | undefined;
 
   private storageManager: StorageManager | undefined;
@@ -367,6 +371,9 @@ export class Matterbridge extends EventEmitter {
       - reset:                 remove the commissioning for Matterbridge (bridge mode). Shutdown Matterbridge before using it!
       - factoryreset:          remove all commissioning information and reset all internal storages. Shutdown Matterbridge before using it!
       - list:                  list the registered plugins
+      - loginterfaces:         log the network interfaces
+      - logstorage:            log the node storage
+      - ssl:                   enable SSL for the frontend and WebSockerServer (certificates in .matterbridge/certs directory cert.pem, key.pem and ca.pem (optional))
       - add [plugin path]:     register the plugin from the given absolute or relative path
       - add [plugin name]:     register the globally installed plugin with the given name
       - remove [plugin path]:  remove the plugin from the given absolute or relative path
@@ -920,7 +927,7 @@ export class Matterbridge extends EventEmitter {
       this.checkUpdateInterval && clearInterval(this.checkUpdateInterval);
       this.checkUpdateInterval = undefined;
 
-      // Calling the shutdown functions with a reason
+      // Calling the shutdown method of each plugin
       for (const plugin of this.registeredPlugins) {
         if (!plugin.enabled || plugin.error) continue;
         this.log.info(`Shutting down plugin ${plg}${plugin.name}${nf}`);
@@ -978,14 +985,29 @@ export class Matterbridge extends EventEmitter {
       // Close the express server
       if (this.expressServer) {
         this.expressServer.close();
+        this.expressServer.removeAllListeners();
         this.expressServer = undefined;
         this.log.debug('Express server closed successfully');
       }
-      // Remove listeners
+      // Close the http server
+      if (this.httpServer) {
+        this.httpServer.close();
+        this.httpServer.removeAllListeners();
+        this.httpServer = undefined;
+        this.log.debug('Frontend http server closed successfully');
+      }
+      // Close the https server
+      if (this.httpsServer) {
+        this.httpsServer.close();
+        this.httpsServer.removeAllListeners();
+        this.httpsServer = undefined;
+        this.log.debug('Frontend https server closed successfully');
+      }
+      // Remove listeners from the express app
       if (this.expressApp) {
         this.expressApp.removeAllListeners();
         this.expressApp = undefined;
-        this.log.debug('Frontend closed successfully');
+        this.log.debug('Frontend app closed successfully');
       }
       // Close the WebSocket server
       if (this.webSocketServer) {
@@ -2431,25 +2453,28 @@ export class Matterbridge extends EventEmitter {
   private async logNodeAndSystemInfo() {
     // IP address information
     const networkInterfaces = os.networkInterfaces();
-    this.systemInformation.ipv4Address = 'Not found';
-    this.systemInformation.ipv6Address = 'Not found';
+    this.systemInformation.ipv4Address = '';
+    this.systemInformation.ipv6Address = '';
     for (const [interfaceName, interfaceDetails] of Object.entries(networkInterfaces)) {
       if (!interfaceDetails) {
         break;
       }
       for (const detail of interfaceDetails) {
-        if (detail.family === 'IPv4' && !detail.internal && this.systemInformation.ipv4Address === 'Not found') {
+        if (detail.family === 'IPv4' && !detail.internal && this.systemInformation.ipv4Address === '') {
           this.systemInformation.interfaceName = interfaceName;
           this.systemInformation.ipv4Address = detail.address;
           this.systemInformation.macAddress = detail.mac;
-        } else if (detail.family === 'IPv6' && !detail.internal && this.systemInformation.ipv6Address === 'Not found') {
+        } else if (detail.family === 'IPv6' && !detail.internal && this.systemInformation.ipv6Address === '') {
           this.systemInformation.interfaceName = interfaceName;
           this.systemInformation.ipv6Address = detail.address;
           this.systemInformation.macAddress = detail.mac;
         }
       }
-      // Break if both addresses are found to improve efficiency
-      if (this.systemInformation.ipv4Address !== 'Not found' && this.systemInformation.ipv6Address !== 'Not found') {
+      if (this.systemInformation.ipv4Address !== '' /* && this.systemInformation.ipv6Address !== ''*/) {
+        this.log.debug(`Using interface: '${this.systemInformation.interfaceName}'`);
+        this.log.debug(`- with MAC address: '${this.systemInformation.macAddress}'`);
+        this.log.debug(`- with IPv4 address: '${this.systemInformation.ipv4Address}'`);
+        this.log.debug(`- with IPv6 address: '${this.systemInformation.ipv6Address}'`);
         break;
       }
     }
@@ -2800,33 +2825,107 @@ export class Matterbridge extends EventEmitter {
    * @param port The port number to run the frontend server on. Default is 3000.
    */
   async initializeFrontend(port = 8283): Promise<void> {
-    if (hasParameter('test_https')) {
-      await this.initializeHttpsFrontend(8443);
-      return;
+    this.log.debug(`Initializing the frontend ${hasParameter('ssl') ? 'https' : 'http'} server on port ${YELLOW}${port}${db}`);
+
+    // Create the express app that serves the frontend
+    this.expressApp = express();
+    this.expressApp.use(express.static(path.join(this.rootDirectory, 'frontend/build')));
+
+    if (!hasParameter('ssl')) {
+      // Create an HTTP server and attach the express app
+      this.httpServer = createServer(this.expressApp);
+
+      // Listen on the specified port
+      this.httpServer.listen(port, () => {
+        if (this.systemInformation.ipv4Address !== '') this.log.info(`The frontend http server is listening on ${UNDERLINE}http://${this.systemInformation.ipv4Address}:${port}${UNDERLINEOFF}${rs}`);
+        if (this.systemInformation.ipv6Address !== '') this.log.debug(`The frontend http server is listening on ${UNDERLINE}http://[${this.systemInformation.ipv6Address}]:${port}${UNDERLINEOFF}${rs}`);
+      });
+
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      this.httpServer.on('error', (error: any) => {
+        this.log.error(`Frontend http server error listening on ${port}`);
+        switch (error.code) {
+          case 'EACCES':
+            this.log.error(`Port ${port} requires elevated privileges`);
+            break;
+          case 'EADDRINUSE':
+            this.log.error(`Port ${port} is already in use`);
+            break;
+        }
+        process.exit(1);
+      });
+    } else {
+      // Load the SSL certificate, the private key and optionally the CA certificate
+      let cert: string | undefined;
+      try {
+        cert = await fs.readFile(path.join(this.matterbridgeDirectory, 'certs/cert.pem'), 'utf8');
+        this.log.info(`Loaded certificate file ${path.join(this.matterbridgeDirectory, 'certs/cert.pem')}`);
+      } catch (error) {
+        this.log.error(`Error reading certificate file: ${error}`);
+        process.exit(1);
+      }
+      let key: string | undefined;
+      try {
+        key = await fs.readFile(path.join(this.matterbridgeDirectory, 'certs/key.pem'), 'utf8');
+        this.log.info(`Loaded key file ${path.join(this.matterbridgeDirectory, 'certs/key.pem')}`);
+      } catch (error) {
+        this.log.error(`Error reading key file: ${error}`);
+        process.exit(1);
+      }
+      let ca: string | undefined;
+      try {
+        ca = await fs.readFile(path.join(this.matterbridgeDirectory, 'certs/ca.pem'), 'utf8');
+        this.log.info(`Loaded CA certificate file ${path.join(this.matterbridgeDirectory, 'certs/ca.pem')}`);
+      } catch (error) {
+        this.log.info(`CA certificate file ${path.join(this.matterbridgeDirectory, 'certs/ca.pem')} not loaded: ${error}`);
+      }
+      const serverOptions: https.ServerOptions = { cert, key, ca };
+
+      // Create an HTTPS server with the SSL certificate and private key (ca is optional) and attach the express app
+      this.httpsServer = https.createServer(serverOptions, this.expressApp);
+
+      // Listen on the specified port
+      this.httpsServer.listen(port, () => {
+        if (this.systemInformation.ipv4Address !== '') this.log.info(`The frontend https server is listening on ${UNDERLINE}https://${this.systemInformation.ipv4Address}:${port}${UNDERLINEOFF}${rs}`);
+        if (this.systemInformation.ipv6Address !== '') this.log.debug(`The frontend https server is listening on ${UNDERLINE}https://[${this.systemInformation.ipv6Address}]:${port}${UNDERLINEOFF}${rs}`);
+      });
+
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      this.httpsServer.on('error', (error: any) => {
+        this.log.error(`Frontend https server error listening on ${port}`);
+        switch (error.code) {
+          case 'EACCES':
+            this.log.error(`Port ${port} requires elevated privileges`);
+            break;
+          case 'EADDRINUSE':
+            this.log.error(`Port ${port} is already in use`);
+            break;
+        }
+        process.exit(1);
+      });
     }
-    this.log.debug(`Initializing the frontend on port ${YELLOW}${port}${db} static ${UNDERLINE}${path.join(this.rootDirectory, 'frontend/build')}${UNDERLINEOFF}${rs}`);
 
-    // Create a WebSocket server
-    const wssPort = 8284;
-    const wssHost = `ws://${this.systemInformation.ipv4Address}:${wssPort}`;
-    this.webSocketServer = new WebSocketServer({ port: wssPort, host: this.systemInformation.ipv4Address });
-    this.log.debug(`WebSocket server created on ${UNDERLINE}${wssHost}${UNDERLINEOFF}${rs}`);
+    // Createe a WebSocket server and attach it to the http server
+    const wssPort = port;
+    const wssHost = hasParameter('ssl') ? `wss://${this.systemInformation.ipv4Address}:${wssPort}` : `ws://${this.systemInformation.ipv4Address}:${wssPort}`;
+    this.webSocketServer = new WebSocketServer(hasParameter('ssl') ? { server: this.httpsServer } : { server: this.httpServer });
 
-    this.webSocketServer.on('connection', (ws: WebSocket) => {
-      this.log.info('WebSocketServer client connected');
+    this.webSocketServer.on('connection', (ws: WebSocket, request: http.IncomingMessage) => {
+      const clientIp = request.socket.remoteAddress;
+      this.log.info(`WebSocketServer client ${clientIp} connected`);
       this.log.setGlobalCallback(this.wssSendMessage.bind(this));
       this.log.debug('WebSocketServer logger callback added');
       this.wssSendMessage('Matterbridge', 'info', 'WebSocketServer client connected to Matterbridge');
 
       ws.on('message', (message) => {
-        this.log.info(`WebSocket client sent a message => ${message}`);
+        this.log.debug(`WebSocket client message: ${message}`);
       });
 
       ws.on('close', () => {
         this.log.info('WebSocket client disconnected');
         if (this.webSocketServer?.clients.size === 0) {
           this.log.setGlobalCallback(undefined);
-          this.log.debug('All WebSocket client disconnected. WebSocketServer logger callback removed');
+          this.log.debug('All WebSocket clients disconnected. WebSocketServer logger callback removed');
         }
       });
 
@@ -2835,16 +2934,32 @@ export class Matterbridge extends EventEmitter {
       });
     });
 
+    this.webSocketServer.on('close', () => {
+      this.log.debug(`WebSocketServer closed`);
+    });
+
+    this.webSocketServer.on('listening', () => {
+      this.log.info(`The WebSocketServer is listening on ${UNDERLINE}${wssHost}${UNDERLINEOFF}${rs}`);
+    });
+
     this.webSocketServer.on('error', (ws: WebSocket, error: Error) => {
       this.log.error(`WebSocketServer error: ${error}`);
-      return;
     });
+
+    /*
+    // Create a WebSocket server
+    const wssPort = 8284;
+    const wssHost = `ws://${this.systemInformation.ipv4Address}:${wssPort}`;
+    this.webSocketServer = new WebSocketServer({ port: wssPort, host: this.systemInformation.ipv4Address });
+    this.log.debug(`WebSocket server created on ${UNDERLINE}${wssHost}${UNDERLINEOFF}${rs}`);
 
     this.webSocketServer.on('listening', () => {
       this.log.info(`WebSocketServer is listening on ${UNDERLINE}${wssHost}${UNDERLINEOFF}${rs}`);
       return;
     });
+    */
 
+    /*
     // Serve React build directory
     this.expressApp = express();
     this.expressApp.use(express.static(path.join(this.rootDirectory, 'frontend/build')));
@@ -2868,6 +2983,7 @@ export class Matterbridge extends EventEmitter {
       }
       process.exit(1);
     });
+    */
 
     // Endpoint to validate login code
     this.expressApp.post('/api/login', express.json(), async (req, res) => {
@@ -3260,80 +3376,14 @@ export class Matterbridge extends EventEmitter {
       res.json({ message: 'Command received' });
     });
 
-    // Fallback for routing (must be the last route)
+    // Fallback for routing (must be the last route but it should not be used because the frontend is static)
     this.expressApp.get('*', (req, res) => {
-      this.log.debug('The frontend sent:', req.url);
-      this.log.debug('Response send file:', path.join(this.rootDirectory, 'frontend/build/index.html'));
+      this.log.warn('The frontend sent:', req.url);
+      this.log.warn('Response send file:', path.join(this.rootDirectory, 'frontend/build/index.html'));
       res.sendFile(path.join(this.rootDirectory, 'frontend/build/index.html'));
     });
 
     this.log.debug(`Frontend initialized on port ${YELLOW}${port}${db} static ${UNDERLINE}${path.join(this.rootDirectory, 'frontend/build')}${UNDERLINEOFF}${rs}`);
-  }
-
-  // Just for testing purposes. Use with matterbridge -test_https so no bridge is loaded
-  async initializeHttpsFrontend(port = 8443): Promise<void> {
-    this.log.debug(`Initializing the https frontend on port ${YELLOW}${port}${db} static ${UNDERLINE}${path.join(this.rootDirectory, 'frontend/build')}${UNDERLINEOFF}${rs}`);
-
-    // Create the express app that serves the frontend
-    this.expressApp = express();
-
-    // Load the SSL certificate and private key
-    // If we need also the CA certificate, we can add it to the serverOptions object
-    // I created dev certificates with openssl using the following commands:
-    // openssl genrsa -out key.pem 2048
-    // openssl req -new -key key.pem -out csr.pem
-    // openssl x509 -req -days 365 -in csr.pem -signkey key.pem -out cert.pem
-    const serverOptions: https.ServerOptions = {
-      cert: await fs.readFile(path.join(this.matterbridgeDirectory, 'certs/cert.pem'), 'utf8'),
-      key: await fs.readFile(path.join(this.matterbridgeDirectory, 'certs/key.pem'), 'utf8'),
-      // ca: await fs.readFile(path.join(this.matterbridgeDirectory, 'certs/ca.pem'), 'utf8'),
-    };
-
-    // Create an HTTPS server with the SSL certificate and private key and attach the express app
-    const httpsServer = https.createServer(serverOptions, this.expressApp);
-
-    // Fallback for routing
-    this.expressApp.get('/', (req, res) => {
-      this.log.debug('The frontend sent:', req.url);
-      res.send('Hello, HTTPS!');
-    });
-
-    // Listen on a specific port
-    // Connect to the frontend with 'https://localhost:8443'
-    httpsServer.listen(port, () => {
-      this.log.info(`HTTPS server listening on ${UNDERLINE}'https://localhost:${port}${UNDERLINEOFF}${rs}`);
-    });
-
-    // Attach WebSocket server to HTTPS server
-    this.webSocketServer = new WebSocketServer({ server: httpsServer });
-    this.log.info(`WebSocketServer runnig on ${UNDERLINE}wss://localhost:${port}${UNDERLINEOFF}${rs}`);
-
-    // Handle connections to the WebSocketServer
-    // Connect to the WebSocket server using wss from the react frontend this way: const ws = new WebSocket('wss://localhost:8443');
-    this.webSocketServer.on('connection', (ws: WebSocket) => {
-      this.log.info('WebSocketServer: WebSocket client connected');
-
-      ws.on('message', (message) => {
-        this.log.info(`WebSocket client sent message => ${message}`);
-      });
-
-      ws.on('close', () => {
-        this.log.info('WebSocket client disconnected');
-        if (this.webSocketServer?.clients.size === 0) {
-          this.log.info('All WebSocket clients disconnected');
-        }
-      });
-
-      ws.on('error', (error: Error) => {
-        this.log.error(`WebSocket client error: ${error}`);
-      });
-    });
-
-    // Handle error from the WebSocketServer
-    this.webSocketServer.on('error', (ws: WebSocket, error: Error) => {
-      this.log.error(`WebSocketServer error: ${error}`);
-      return;
-    });
   }
 
   /**