+++ src/utils/export.ts
@@ -1,2 +1,0 @@
-export * from './utils.js';
-export * from './colorUtils.js';